requirements.txt
streamlit 
langchain 
langchain-core 
langchain-community 
langchain-groq 
python-dotenv
langchain-tavily